Fix Image editor, mimetype, size validation errors, unable to save, and more

+++ packages/forms/resources/js/components/file-upload.js
@@ -40,6 +40,7 @@ export default function fileUploadFormComponent({
     isDeletable,
     isDisabled,
     isDownloadable,
+    isMultiple,
     isOpenable,
     isPreviewable,
     isReorderable,
@@ -500,24 +501,44 @@ export default function fileUploadFormComponent({
                     width: imageResizeTargetWidth,
                 })
                 .toBlob((croppedImage) => {
-                    this.pond.removeFile(
-                        this.pond
-                            .getFiles()
-                            .find(
-                                (uploadedFile) =>
-                                    uploadedFile.filename ===
-                                    this.editingFile.name,
-                            ),
-                    )
+                    if (isMultiple) {
+                        this.pond.removeFile(
+                            this.pond
+                                .getFiles()
+                                .find(
+                                    (uploadedFile) =>
+                                        uploadedFile.filename ===
+                                        this.editingFile.name,
+                                )?.id,
+                            { revert: true },
+                        )
+                    }
 
                     this.$nextTick(() => {
                         this.shouldUpdateState = false
+
+                        let editingFileName = this.editingFile.name.slice(0, this.editingFile.name.lastIndexOf('.'))
+                        let editingFileExtension = this.editingFile.name.slice(this.editingFile.name.lastIndexOf('.') + 1)
+
+                        let editingFileNameNumber = null
+
+                        if (editingFileName.includes('-')) {
+                            editingFileNameNumber = +editingFileName.slice(editingFileName.lastIndexOf('-') + 1)
+                        }
+
+                        let newEditingFileName;
+
+                        if ((editingFileNameNumber !== null) && (! isNaN(editingFileNameNumber))) {
+                            newEditingFileName = `${editingFileName.slice(0, editingFileName.lastIndexOf('-')).substring(0, 64)}-${Math.floor(Math.random() * 100000)}.${editingFileExtension}`
+                        }
+
+                        newEditingFileName ??= `${editingFileName.substring(0, 64)}-${Math.floor(Math.random() * 100000)}.${editingFileExtension}`
 
                         this.pond
                             .addFile(
                                 new File(
                                     [croppedImage],
-                                    this.editingFile.name,
+                                    newEditingFileName,
                                     {
                                         type:
                                             this.editingFile.type ===
